fix(AB#39533): minor improvements, move to utils
