+++ packages/react/src/select/select-next.tsx
@@ -16,7 +16,8 @@ import { useScrollHighlightedItem } from '@/hooks/use-scroll-highlighted-item.js
 import { translate as t } from '@/i18n/utility.js';
 import { InputText } from '@/input-text/input-text.js';
 import { Popover } from '@/popover/popover.js';
-import { cycleEnabledIndex, getTextContent } from '@/utilities.js';
+import { cycleEnabledIndex } from '@/utilities.js';
+import { getTextContent } from '@/utils/utilities.js';
 import { SelectMenu, SelectMenuGroupItem, SelectMenuOption } from './select-menu.js';
 import { SelectSearch } from './select-search.js';
 import type {
